Reorganize mocha test setup
